LayoutTests/accessibility/self-referencing-aria-labelledby-expected.txt

This tests that self-referencing aria-labelledby is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS platformValueForW3CName(test) is "X Y Z"
PASS successfullyParsed is true

TEST COMPLETE
